Addresses some PR feedback
